cell-poc/cell-1/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
prometheus-client==0.19.0